Relocate getRelativeKeyTonic
